server/requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
paho-mqtt==1.6.1
eventlet==0.33.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
